Add separate methods for changing station and vehicle labels
This commit introduces new methods for changing just the label of a station or a vehicle. Previously, modification of label and group of a station or a vehicle was available only through a single method. This helps to offer more granularity to the users who can now choose to modify either just the label or both label and the group.